pyproject.toml
[tool.poetry]
name = "hh-coursework-4"
version = "0.1.0"
description = ""
authors = ["TheShadowfish <shadowfish@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"